remove redundant database by integrating wip into main db
