+++ src/llm/config.dev.yaml
@@ -1,0 +1,79 @@
+# Router group aliases: optional per-model-row ``model_name`` assigns a deployment to an alias
+# (default for ``providers`` is ``default_model_name``; for ``fallback_providers`` it is
+# ``fallback_model_name``). Use ``_call(..., model="writer")`` etc. to target an alias.
+#
+# Example (commented):
+#   models:
+#     - model: "gemini/gemini-2.5-pro"
+#       model_name: "app"
+#     - model: "gemini/gemini-2.5-flash"
+#       model_name: "writer"
+
+router:
+  default_model_name: "app"
+  fallback_model_name: "fallback"
+  fallbacks:
+    - app: ["fallback"]
+    # Planner has its own alias chain so it can prefer only the largest models.
+    - planner: ["planner_openrouter_large"]
+    - planner_openrouter_large: ["planner_gemini_fallback"]
+    # Slide writing stays Gemini-first, then falls back to the generic pool.
+    - slides: ["fallback"]
+    # - writer: ["fallback"]
+    # - fast: ["fallback"]
+
+  providers:
+    gemini:
+      api_key: "os.environ/GEMINI_API_KEY"
+      models:
+        - model: "gemini/gemini-3.1-flash-lite-preview"
+          model_name: "planner_gemini_fallback"
+          rpm: 15
+        - model: "gemini/gemini-3.1-flash-lite-preview"
+          model_name: "slides"
+          rpm: 15
+
+  # Note: splitting fallback_providers into *_high and *_rest allows us to control the 
+  # exact model-level fallback order across different providers (e.g. OpenRouter -> Ollama -> OpenRouter)
+  # since the Router follows the order entries appear in this YAML.
+  fallback_providers:
+    openrouter_high:
+      api_key: "os.environ/OPENROUTER_API_KEY"
+      api_base: "os.environ/OPENROUTER_BASE_URL"
+      models:
+        - model: "openrouter/openai/gpt-oss-120b:free"
+          model_name: "planner_openrouter_large"
+        - model: "openrouter/openai/gpt-oss-120b:free"
+
+    ollama_high:
+      api_key: "os.environ/OLLAMA_API_KEY"
+      api_base: "os.environ/OLLAMA_BASE_URL"
+      models:
+        - model: "ollama/qwen3.5:397b-cloud"
+          model_name: "planner"
+        - model: "ollama/qwen3.5:397b-cloud"
+
+    gemini_rest:
+      api_key: "os.environ/GEMINI_API_KEY"
+      models:
+        - model: "gemini/gemma-4-31b-it"
+          rpm: 15
+
+    ollama_rest:
+      api_key: "os.environ/OLLAMA_API_KEY"
+      api_base: "os.environ/OLLAMA_BASE_URL"
+      models:
+        - model: "ollama/gemma4:31b-cloud"
+
+    openrouter_rest:
+      api_key: "os.environ/OPENROUTER_API_KEY"
+      api_base: "os.environ/OPENROUTER_BASE_URL"
+      models:
+        - model: "openrouter/openai/gpt-oss-20b:free"
+
+  settings:
+    num_retries: 2
+    retry_after: 5
+    cooldown_time: 30
+    enable_pre_call_checks: true
+    timeout: 120
