--- src/memory/wip/__init__.py
@@ -1,11 +1,0 @@
-from .schema import ProtoSlide, SlideContent
-from .database import WIPDatabase
-from .tools import read_proto_slide, write_proto_slide
-
-__all__ = [
-    "ProtoSlide",
-    "SlideContent",
-    "WIPDatabase",
-    "read_proto_slide",
-    "write_proto_slide"
-]
